requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.6.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
boto3==1.28.73
botocore==1.31.73
bs4==0.0.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.6.1
deprecation==2.1.0
distlib==0.3.7
distro==1.8.0
dnspython==2.4.2
fastapi==0.95.1
filelock==3.13.0
Flask==3.0.0
frozenlist==1.3.3
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
gotrue==1.3.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
httpx-sse==0.3.1
idna==3.4
importlib-metadata==6.6.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.330
langcodes==3.3.0
langserve==0.0.24
langsmith==0.0.60
loguru==0.7.2
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numpy==1.24.3
openai==1.3.8
packaging==23.2
pandas==2.1.3
pinecone-client==2.2.4
pipenv==2023.10.24
platformdirs==3.11.0
postgrest==0.13.0
preshed==3.0.9
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pydantic==1.10.7
pydantic_core==2.10.1
Pygments==2.16.1
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywatchman==1.4.1
realtime==1.0.0
requests==2.30.0
rich==13.6.0
rsa==4.9
s3transfer==0.7.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
srsly==2.4.8
sse-starlette==1.6.5
starlette==0.26.1
storage3==0.6.1
StrEnum==0.4.15
tenacity==8.2.3
textstat==0.7.3
thinc==8.2.1
tomli==2.0.1
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.6.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.21.1
virtualenv==20.24.6
wasabi==1.1.2
weasel==0.3.4
websockets==10.4
yarl==1.9.2
zipp==3.15.0
pypdf==3.17.1
python-dotenv==1.0.0
asyncio==3.4.3